alter price field

+++ apps/dashboard/forms/price.py
@@ -53,7 +53,7 @@ class PriceForm(ModelForm):
             },
             format="%H:%M"
         ))
-    full_day_price = forms.FloatField(
+    full_day_price = forms.DecimalField(
         required=False,
         widget=forms.TextInput(
             attrs={
